+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getUSDCMoonpayWarpConfig.ts
@@ -5,6 +5,7 @@ import {
   HypTokenRouterConfig,
   IsmType,
   OwnableConfig,
+  RoutingIsmConfig,
   TokenType,
 } from '@hyperlane-xyz/sdk';
 
@@ -15,8 +16,8 @@ import {
   tokens,
 } from '../../../../../src/config/warp.js';
 import { getExistingWarpDeployConfig } from './utils.js';
-import { WarpRouteIds } from '../warpIds.js';
 
+const STAGING_SOURCE_WARP_ROUTE_ID = 'USDC/ctusd';
 const FAST_PATH_RELAYER = relayerAddresses.mainnet3.fastpath;
 const HYPERLANE_RELAYER = relayerAddresses.mainnet3.hyperlane;
 const CCTP_CAPABLE_CHAINS: ChainName[] = ['arbitrum', 'base', 'ethereum'];
@@ -24,9 +25,14 @@ const ROUTE_CHAINS: ChainName[] = [
   'solanamainnet',
   'arbitrum',
   'base',
-  'ethereum',
   'citrea',
+  'ethereum',
+  'katana',
 ];
+const SOLANA_XO_TOKEN_MINT = 'xoUSDq85Rjsb6SbUwJyreFgeWQvxdkT7R3c3g7s6p5Y';
+const SOLANA_XO_NAME = 'XO Cash';
+const SOLANA_XO_SYMBOL = 'XO';
+const KATANA_VBUSDC_TOKEN = '0x203A662b0BD271A6ed5a60EdFbd04bFce608FD36';
 
 function trustedRelayerIsm(relayer: string) {
   return {
@@ -59,31 +65,36 @@ function getTrustedRelayer(local: ChainName, remote: ChainName): string {
     : FAST_PATH_RELAYER;
 }
 
-function getInterchainSecurityModule(local: ChainName, owner: string) {
+function getInterchainSecurityModule(
+  local: ChainName,
+  owner: string,
+): RoutingIsmConfig | ReturnType<typeof trustedRelayerIsm> {
   if (local === 'solanamainnet') {
     return trustedRelayerIsm(HYPERLANE_RELAYER);
   }
 
+  const domains = Object.fromEntries(
+    ROUTE_CHAINS.filter((chain) => chain !== local).map((remote) => [
+      remote,
+      local !== 'citrea' && CCTP_CAPABLE_CHAINS.includes(remote)
+        ? cctpOrTrustedRelayerIsm(owner, getTrustedRelayer(local, remote))
+        : trustedRelayerIsm(getTrustedRelayer(local, remote)),
+    ]),
+  );
+
   return {
     type: IsmType.ROUTING,
     owner,
-    domains: Object.fromEntries(
-      ROUTE_CHAINS.filter((chain) => chain !== local).map((remote) => [
-        remote,
-        CCTP_CAPABLE_CHAINS.includes(remote)
-          ? cctpOrTrustedRelayerIsm(owner, getTrustedRelayer(local, remote))
-          : trustedRelayerIsm(getTrustedRelayer(local, remote)),
-      ]),
-    ),
-  } as const;
+    domains,
+  };
 }
 
-export async function getUSDTCtUSDWarpConfig(
+export async function getUSDCMoonpayWarpConfig(
   routerConfig: ChainMap<RouterConfigWithoutOwner>,
   abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
 ): Promise<ChainMap<HypTokenRouterConfig>> {
   const existingConfig = await getExistingWarpDeployConfig(
-    WarpRouteIds.USDTCtUSD,
+    STAGING_SOURCE_WARP_ROUTE_ID,
   );
   const solanamainnetOwner =
     existingConfig.solanamainnet.owner ??
@@ -92,6 +103,12 @@ export async function getUSDTCtUSDWarpConfig(
     existingConfig.arbitrum.owner ?? abacusWorksEnvOwnerConfig.arbitrum.owner;
   const baseOwner =
     existingConfig.base.owner ?? abacusWorksEnvOwnerConfig.base.owner;
+  const citreaOwner =
+    existingConfig.citrea.owner ?? abacusWorksEnvOwnerConfig.citrea.owner;
+  const ethereumOwner =
+    existingConfig.ethereum.owner ?? abacusWorksEnvOwnerConfig.ethereum.owner;
+  const katanaOwner =
+    existingConfig.katana.owner ?? abacusWorksEnvOwnerConfig.katana.owner;
 
   return {
     ...existingConfig,
@@ -100,7 +117,9 @@ export async function getUSDTCtUSDWarpConfig(
       mailbox: routerConfig.solanamainnet.mailbox,
       owner: solanamainnetOwner,
       type: TokenType.crossCollateral,
-      token: tokens.solanamainnet.USDT,
+      token: SOLANA_XO_TOKEN_MINT,
+      name: SOLANA_XO_NAME,
+      symbol: SOLANA_XO_SYMBOL,
       decimals: 6,
       interchainSecurityModule: getInterchainSecurityModule(
         'solanamainnet',
@@ -112,7 +131,7 @@ export async function getUSDTCtUSDWarpConfig(
       mailbox: routerConfig.arbitrum.mailbox,
       owner: arbitrumOwner,
       type: TokenType.crossCollateral,
-      token: tokens.arbitrum.USDT,
+      token: tokens.arbitrum.USDC,
       interchainSecurityModule: getInterchainSecurityModule(
         'arbitrum',
         arbitrumOwner,
@@ -123,8 +142,41 @@ export async function getUSDTCtUSDWarpConfig(
       mailbox: routerConfig.base.mailbox,
       owner: baseOwner,
       type: TokenType.crossCollateral,
-      token: tokens.base.USDT,
+      token: tokens.base.USDC,
       interchainSecurityModule: getInterchainSecurityModule('base', baseOwner),
+    },
+    citrea: {
+      ...existingConfig.citrea,
+      mailbox: routerConfig.citrea.mailbox,
+      owner: citreaOwner,
+      type: TokenType.crossCollateral,
+      token: tokens.citrea.ctUSD,
+      interchainSecurityModule: getInterchainSecurityModule(
+        'citrea',
+        citreaOwner,
+      ),
+    },
+    ethereum: {
+      ...existingConfig.ethereum,
+      mailbox: routerConfig.ethereum.mailbox,
+      owner: ethereumOwner,
+      type: TokenType.crossCollateral,
+      token: tokens.ethereum.USDC,
+      interchainSecurityModule: getInterchainSecurityModule(
+        'ethereum',
+        ethereumOwner,
+      ),
+    },
+    katana: {
+      ...existingConfig.katana,
+      mailbox: routerConfig.katana.mailbox,
+      owner: katanaOwner,
+      type: TokenType.crossCollateral,
+      token: KATANA_VBUSDC_TOKEN,
+      interchainSecurityModule: getInterchainSecurityModule(
+        'katana',
+        katanaOwner,
+      ),
     },
   };
 }
