+++ typescript/infra/config/environments/mainnet3/warp/warpIds.ts
@@ -116,8 +116,8 @@ export enum WarpRouteIds {
   MainnetCCTPV1 = 'USDC/mainnet-cctp',
   MainnetCCTPV2Fast = 'USDC/mainnet-cctp-v2-fast',
   MainnetCCTPV2Standard = 'USDC/mainnet-cctp-v2-standard',
-  USDCCtUSD = 'USDC/ctusd',
-  USDTCtUSD = 'USDT/ctusd',
+  USDCMoonpay = 'USDC/moonpay',
+  USDTMoonpay = 'USDT/moonpay',
   TestnetCCTPV1 = 'USDC/testnet-cctp',
 
   // HYPER routes
